rename metdada to registries
