+++ annotator-core/src/main/java/edu/ucr/cs/riple/core/checkers/Checker.java
@@ -26,11 +26,11 @@ package edu.ucr.cs.riple.core.checkers;
 
 import com.google.common.collect.ImmutableSet;
 import edu.ucr.cs.riple.core.injectors.AnnotationInjector;
-import edu.ucr.cs.riple.core.metadata.index.Error;
-import edu.ucr.cs.riple.core.metadata.index.Fix;
-import edu.ucr.cs.riple.core.metadata.region.Region;
 import edu.ucr.cs.riple.core.module.ModuleConfiguration;
 import edu.ucr.cs.riple.core.module.ModuleInfo;
+import edu.ucr.cs.riple.core.registries.index.Error;
+import edu.ucr.cs.riple.core.registries.index.Fix;
+import edu.ucr.cs.riple.core.registries.region.Region;
 import java.util.Set;
 
 /**
