+++ annotator-core/src/main/java/edu/ucr/cs/riple/core/checkers/nullaway/NullAwayError.java
@@ -25,9 +25,9 @@
 package edu.ucr.cs.riple.core.checkers.nullaway;
 
 import com.google.common.collect.ImmutableSet;
-import edu.ucr.cs.riple.core.metadata.index.Error;
-import edu.ucr.cs.riple.core.metadata.index.Fix;
-import edu.ucr.cs.riple.core.metadata.region.Region;
+import edu.ucr.cs.riple.core.registries.index.Error;
+import edu.ucr.cs.riple.core.registries.index.Fix;
+import edu.ucr.cs.riple.core.registries.region.Region;
 import java.util.Objects;
 
 /** Represents an error reported by {@link NullAway}. */
