+++ annotator-core/src/main/java/edu/ucr/cs/riple/core/evaluators/CachedEvaluator.java
@@ -33,8 +33,8 @@ import edu.ucr.cs.riple.core.Report;
 import edu.ucr.cs.riple.core.cache.Impact;
 import edu.ucr.cs.riple.core.cache.TargetModuleCache;
 import edu.ucr.cs.riple.core.evaluators.suppliers.Supplier;
-import edu.ucr.cs.riple.core.metadata.index.Error;
-import edu.ucr.cs.riple.core.metadata.index.Fix;
+import edu.ucr.cs.riple.core.registries.index.Error;
+import edu.ucr.cs.riple.core.registries.index.Fix;
 import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
