+++ annotator-core/src/main/java/edu/ucr/cs/riple/core/evaluators/graph/processors/ParallelConflictGraphProcessor.java
@@ -28,11 +28,11 @@ import edu.ucr.cs.riple.core.Context;
 import edu.ucr.cs.riple.core.evaluators.graph.ConflictGraph;
 import edu.ucr.cs.riple.core.evaluators.graph.Node;
 import edu.ucr.cs.riple.core.evaluators.suppliers.Supplier;
-import edu.ucr.cs.riple.core.metadata.index.Error;
-import edu.ucr.cs.riple.core.metadata.index.Fix;
-import edu.ucr.cs.riple.core.metadata.index.Result;
-import edu.ucr.cs.riple.core.metadata.region.Region;
-import edu.ucr.cs.riple.core.metadata.region.RegionRegistry;
+import edu.ucr.cs.riple.core.registries.index.Error;
+import edu.ucr.cs.riple.core.registries.index.Fix;
+import edu.ucr.cs.riple.core.registries.index.Result;
+import edu.ucr.cs.riple.core.registries.region.Region;
+import edu.ucr.cs.riple.core.registries.region.RegionRegistry;
 import edu.ucr.cs.riple.core.util.Utility;
 import java.util.Collection;
 import java.util.HashSet;
