+++ annotator-core/src/main/java/edu/ucr/cs/riple/core/evaluators/graph/processors/SequentialConflictGraphProcessor.java
@@ -27,8 +27,8 @@ package edu.ucr.cs.riple.core.evaluators.graph.processors;
 import edu.ucr.cs.riple.core.Context;
 import edu.ucr.cs.riple.core.evaluators.graph.ConflictGraph;
 import edu.ucr.cs.riple.core.evaluators.suppliers.Supplier;
-import edu.ucr.cs.riple.core.metadata.index.Fix;
-import edu.ucr.cs.riple.core.metadata.index.Result;
+import edu.ucr.cs.riple.core.registries.index.Fix;
+import edu.ucr.cs.riple.core.registries.index.Result;
 import edu.ucr.cs.riple.core.util.Utility;
 import java.util.Set;
 import me.tongfei.progressbar.ProgressBar;
