+++ annotator-core/src/main/java/edu/ucr/cs/riple/core/registries/Registry.java
@@ -22,7 +22,7 @@
  * THE SOFTWARE.
  */
 
-package edu.ucr.cs.riple.core.metadata;
+package edu.ucr.cs.riple.core.registries;
 
 import com.google.common.collect.ImmutableMultimap;
 import com.google.common.collect.ImmutableSet;
