+++ annotator-core/src/main/java/edu/ucr/cs/riple/core/registries/region/FieldRegionRegistry.java
@@ -22,30 +22,32 @@
  * THE SOFTWARE.
  */
 
-package edu.ucr.cs.riple.core.metadata.region;
+package edu.ucr.cs.riple.core.registries.region;
 
 import com.google.common.collect.ImmutableSet;
-import edu.ucr.cs.riple.core.metadata.Registry;
-import edu.ucr.cs.riple.core.metadata.method.MethodRecord;
 import edu.ucr.cs.riple.core.module.ModuleInfo;
+import edu.ucr.cs.riple.core.registries.Registry;
 import edu.ucr.cs.riple.core.util.Utility;
 import edu.ucr.cs.riple.injector.location.Location;
-import edu.ucr.cs.riple.injector.location.OnMethod;
+import edu.ucr.cs.riple.injector.location.OnField;
 import edu.ucr.cs.riple.scanner.Serializer;
+import java.util.stream.Collectors;
 
 /**
- * Region registry for Methods. This region registry can identify impacted regions for fixes on
- * {@link OnMethod}
+ * Region registry for Fields. This region registry can identify impacted regions for fixes on
+ * {@link OnField}.
  */
-public class MethodRegionRegistry extends Registry<RegionRecord> implements RegionRegistry {
+public class FieldRegionRegistry extends Registry<RegionRecord> implements RegionRegistry {
 
-  /** ModuleInfo of the module which usage of methods are stored. */
+  /** ModuleInfo of the module which usages of fields are stored. */
   private final ModuleInfo moduleInfo;
 
-  public MethodRegionRegistry(ModuleInfo moduleInfo) {
+  public FieldRegionRegistry(ModuleInfo moduleInfo) {
     super(
         moduleInfo.getModuleConfigurations().stream()
-            .map(info -> info.dir.resolve(Serializer.METHOD_IMPACTED_REGION_FILE_NAME))
+            .map(
+                configuration ->
+                    configuration.dir.resolve(Serializer.FIELD_IMPACTED_REGION_FILE_NAME))
             .collect(ImmutableSet.toImmutableSet()));
     this.moduleInfo = moduleInfo;
   }
@@ -57,36 +59,41 @@ public class MethodRegionRegistry extends Registry<RegionRecord> implements Regi
 
   @Override
   public ImmutableSet<Region> getImpactedRegions(Location location) {
-    if (!location.isOnMethod()) {
+    if (!location.isOnField()) {
       return ImmutableSet.of();
     }
+    OnField field = location.toField();
+    // Add all regions where the field is assigned a new value or read.
     ImmutableSet.Builder<Region> builder = ImmutableSet.builder();
-    OnMethod onMethod = location.toMethod();
-    // Add callers of method.
-    builder.addAll(getImpactedRegionsByUse(onMethod));
-    // Add method itself.
-    builder.add(new Region(onMethod.clazz, onMethod.method));
-    // Add immediate super method.
-    MethodRecord parent = moduleInfo.getMethodRegistry().getImmediateSuperMethod(onMethod);
-    if (parent != null && parent.isNonTop()) {
-      builder.add(new Region(parent.location.clazz, parent.location.method));
+    builder.addAll(getImpactedRegionsByUse(location));
+    // Add each a region for each field variable declared in the declaration statement.
+    builder.addAll(
+        field.variables.stream()
+            .map(fieldName -> new Region(field.clazz, fieldName))
+            .collect(Collectors.toSet()));
+    // Check if field is initialized at declaration.
+    if (moduleInfo.getFieldRegistry().isUninitializedField(field)) {
+      // If not, add all constructors for the class.
+      builder.addAll(
+          moduleInfo.getMethodRegistry().getConstructorsForClass(field.clazz).stream()
+              .map(onMethod -> new Region(onMethod.clazz, onMethod.method))
+              .collect(Collectors.toSet()));
     }
     return builder.build();
   }
 
   @Override
   public ImmutableSet<Region> getImpactedRegionsByUse(Location location) {
-    if (!location.isOnMethod()) {
+    if (!location.isOnField()) {
       return ImmutableSet.of();
     }
-    OnMethod onMethod = location.toMethod();
-    // Add callers of method.
+    OnField field = location.toField();
     return findRecordsWithHashHint(
             candidate ->
-                candidate.calleeClass.equals(onMethod.clazz)
-                    && candidate.calleeMember.equals(onMethod.method),
-            RegionRecord.hash(onMethod.clazz))
-        .map(node -> node.region)
+                candidate.calleeClass.equals(field.clazz)
+                    && field.isOnFieldWithName(candidate.calleeMember),
+            RegionRecord.hash(field.clazz))
+        .map(regionRecord -> regionRecord.region)
         .collect(ImmutableSet.toImmutableSet());
   }
 }
