+++ annotator-core/src/main/java/edu/ucr/cs/riple/core/registries/region/ParameterRegionRegistry.java
@@ -22,7 +22,7 @@
  * THE SOFTWARE.
  */
 
-package edu.ucr.cs.riple.core.metadata.region;
+package edu.ucr.cs.riple.core.registries.region;
 
 import com.google.common.collect.ImmutableSet;
 import edu.ucr.cs.riple.core.module.ModuleInfo;
