+++ annotator-core/src/main/java/edu/ucr/cs/riple/core/registries/region/Region.java
@@ -22,7 +22,7 @@
  * THE SOFTWARE.
  */
 
-package edu.ucr.cs.riple.core.metadata.region;
+package edu.ucr.cs.riple.core.registries.region;
 
 import edu.ucr.cs.riple.injector.Helper;
 import edu.ucr.cs.riple.injector.location.OnClass;
