+++ annotator-core/src/test/java/edu/ucr/cs/riple/core/tools/TError.java
@@ -24,8 +24,8 @@
 
 package edu.ucr.cs.riple.core.tools;
 
-import edu.ucr.cs.riple.core.metadata.index.Error;
-import edu.ucr.cs.riple.core.metadata.region.Region;
+import edu.ucr.cs.riple.core.registries.index.Error;
+import edu.ucr.cs.riple.core.registries.region.Region;
 import edu.ucr.cs.riple.injector.location.Location;
 
 /**
